cfxcore

Module block_data_manager

Modules

block_data_types

db_gc_manager

db_manager

tx_data_manager

Structs

BlamedHeaderVerifiedRoots

Verified roots of blamed headers stored on disk on light nodes.

BlockDataManager

BlockDataWithMultiVersion

The structure to maintain block data under different views.

BlockExecutionResult

receipts and bloom of a single block after execution. It might change depending on this block is executed under which pivot block’s view.

BlockRewardResult

CheckpointHashes

The checkpoint information stored in the database

DataManagerConfiguration

DataVersionTuple

EpochExecutionContext

The start block number of an epoch. It equals to the past executed number of blocks in the previous epoch + 1. For the true genesis, it equals 0. Used in evm execution.

InvalidBlockSet

LocalBlockInfo

The local information about a block. It is NOT consistent across different nodes.

PosRewardForAccount

PosRewardInfo

Enums

BlockStatus

The validity status of a block. If a block’s status among all honest nodes is guaranteed to have no conflict, which means if some honest nodes think a block is not Pending, their decision will be the same status.

DbType

Constants

NULLU64

Functions

db_decode_list

db_encode_list

Type Aliases

BlockExecutionResultWithEpoch

BlockReceiptsInfo

BlockRewardsInfo

BlockTracesInfo

BlockTracesWithEpoch