cfxcore::block_data_manager::db_manager

Struct DBManager

pub struct DBManager { /* private fields */ }

Implementations

impl DBManager

pub fn new_from_rocksdb(db: Arc<SystemDB>, pow: Arc<PowComputer>) -> Self

impl DBManager

pub fn new_from_sqlite(db_path: &Path, pow: Arc<PowComputer>) -> Self

impl DBManager

pub fn insert_block_traces_to_db( &self, block_hash: &H256, block_traces: &BlockTracesWithEpoch, )

pub fn block_traces_from_db( &self, block_hash: &H256, ) -> Option<BlockTracesWithEpoch>

pub fn block_from_db(&self, block_hash: &H256) -> Option<Block>

TODO Use new_with_rlp_size

pub fn insert_block_header_to_db(&self, header: &BlockHeader)

pub fn block_header_from_db(&self, hash: &H256) -> Option<BlockHeader>

pub fn remove_block_header_from_db(&self, hash: &H256)

pub fn insert_transaction_index_to_db( &self, hash: &H256, value: &TransactionIndex, )

pub fn transaction_index_from_db(&self, hash: &H256) -> Option<TransactionIndex>

pub fn insert_hash_by_block_number_to_db(&self, block_number: u64, hash: &H256)

pub fn hash_by_block_number_from_db(&self, block_number: &u64) -> Option<H256>

pub fn insert_local_block_info_to_db( &self, block_hash: &H256, value: &LocalBlockInfo, )

Store block info to db. Block info includes block status and the sequence number when the block enters consensus graph. The db key is the block hash plus one extra byte, so we can get better data locality if we get both a block and its info from db. The info is not a part of the block because the block is inserted before we know its info, and we do not want to insert a large chunk again. TODO Maybe we can use in-place modification (operator merge in rocksdb) to keep the info together with the block.

pub fn local_block_info_from_db( &self, block_hash: &H256, ) -> Option<LocalBlockInfo>

Get block info from db.

pub fn insert_blamed_header_verified_roots_to_db( &self, block_height: u64, value: &BlamedHeaderVerifiedRoots, )

pub fn blamed_header_verified_roots_from_db( &self, block_height: u64, ) -> Option<BlamedHeaderVerifiedRoots>

Get correct roots of blamed headers from db. These are maintained on light nodes only.

pub fn remove_blamed_header_verified_roots_from_db(&self, block_height: u64)

pub fn insert_block_body_to_db(&self, block: &Block)

pub fn block_body_from_db( &self, hash: &H256, ) -> Option<Vec<Arc<SignedTransaction>>>

pub fn remove_block_body_from_db(&self, hash: &H256)

pub fn insert_block_execution_result_to_db( &self, hash: &H256, value: &BlockExecutionResultWithEpoch, )

pub fn insert_block_reward_result_to_db( &self, hash: &H256, value: &DataVersionTuple<H256, BlockRewardResult>, )

pub fn block_execution_result_from_db( &self, hash: &H256, ) -> Option<BlockExecutionResultWithEpoch>

pub fn block_reward_result_from_db( &self, hash: &H256, ) -> Option<DataVersionTuple<H256, BlockRewardResult>>

pub fn remove_block_execution_result_from_db(&self, hash: &H256)

pub fn remove_block_reward_result_from_db(&self, hash: &H256)

pub fn remove_block_trace_from_db(&self, hash: &H256)

pub fn remove_transaction_index_from_db(&self, hash: &H256)

pub fn insert_checkpoint_hashes_to_db( &self, checkpoint_prev: &H256, checkpoint_cur: &H256, )

pub fn checkpoint_hashes_from_db(&self) -> Option<(H256, H256)>

pub fn insert_executed_epoch_set_hashes_to_db( &self, epoch: u64, executed_hashes: &Vec<H256>, )

pub fn insert_skipped_epoch_set_hashes_to_db( &self, epoch: u64, skipped_hashes: &Vec<H256>, )

pub fn executed_epoch_set_hashes_from_db(&self, epoch: u64) -> Option<Vec<H256>>

pub fn skipped_epoch_set_hashes_from_db(&self, epoch: u64) -> Option<Vec<H256>>

pub fn insert_terminals_to_db(&self, terminals: &Vec<H256>)

pub fn terminals_from_db(&self) -> Option<Vec<H256>>

pub fn insert_epoch_execution_commitment_to_db( &self, hash: &H256, ctx: &EpochExecutionCommitment, )

pub fn epoch_execution_commitment_from_db( &self, hash: &H256, ) -> Option<EpochExecutionCommitment>

pub fn remove_epoch_execution_commitment_from_db(&self, hash: &H256)

pub fn insert_instance_id_to_db(&self, instance_id: u64)

pub fn instance_id_from_db(&self) -> Option<u64>

pub fn insert_execution_context_to_db( &self, hash: &H256, ctx: &EpochExecutionContext, )

pub fn execution_context_from_db( &self, hash: &H256, ) -> Option<EpochExecutionContext>

pub fn remove_epoch_execution_context_from_db(&self, hash: &H256)

pub fn insert_gc_progress_to_db(&self, next_to_process: u64)

pub fn gc_progress_from_db(&self) -> Option<u64>

pub fn insert_pos_reward(&self, pos_epoch: u64, pos_reward: &PosRewardInfo)

pub fn pos_reward_by_pos_epoch(&self, pos_epoch: u64) -> Option<PosRewardInfo>