cfxcore::consensus::consensus_inner

Module confirmation_meter

Structs

ConfirmationMeter

ConfirmationMeter computes an approximate local view confirmation risk given the current blockchain state. Local view means that the meter assumes a potential block propagation delay and assumes a worst case scenario of what this delay could do.

FinalityManager

TotalWeightInPastMovingDelta