Module cfxcore::consensus_internal_parameters

Constants

ANTICONE_BARRIER_CAP

This is the cap of the size of the anticone barrier. If we have more than this number we will use the brute_force O(n) algorithm instead.

ANTICONE_PENALTY_RATIO

ANTICONE_PENALTY_UPPER_EPOCH_COUNT

BLOCKSET_IN_OWN_VIEW_OF_EPOCH_CAP

This is the cap of the size of blockset_in_own_view_of_epoch. If we have more than this number, we will not store it in memory

CIP107_STORAGE_POINT_PROP_INIT

The initial storage point proportion after CIP107 is enabled.

CIP137_BASEFEE_PROP_INIT

The initial base price share proportion after CIP137 is enabled.

CONFIRMATION_METER_ADAPTIVE_TEST_TIMER_DIFF

The minimum timer diff value for the adaptive test in confirmation meter to consider

CONFIRMATION_METER_MAXIMUM_ADAPTIVE_RISK

The maximum value of adaptive block generation risk that a confirmation meter is going to consider safe to assume no adaptive blocks in the near future.

CONFIRMATION_METER_MAX_NUM_MAINTAINED_RISK

The maximum number of epochs that the confirmation meter tries to maintain internally.

CONFIRMATION_METER_MIN_MAINTAINED_RISK

This is the minimum risk that the confirmation meter tries to maintain.

CONFIRMATION_METER_PSI

The batch step in the confirmation meter to do the adaptive test

CONFIRMATION_METER_UPDATE_FREQUENCY

This controls how often the confirmation meter updates. The default is to update the meter every 20 blocks. Note that confirmation meter update is CPU intensive if the tree graph is in a unstable state.

DAO_MIN_VOTE_PERCENTAGE

DAO_PARAMETER_VOTE_PERIOD

ELASTICITY_MULTIPLIER

EPOCH_EXECUTED_BLOCK_BOUND

The maximum number of blocks to be executed in each epoch

GENESIS_TOKEN_COUNT_IN_CFX

INITIAL_1559_CORE_BASE_PRICE

The initial and minimum base price

INITIAL_1559_ETH_BASE_PRICE

INITIAL_BASE_MINING_REWARD_IN_UCFX

MINED_BLOCK_COUNT_PER_QUARTER

MINING_REWARD_DECAY_PERIOD_IN_QUARTER

MINING_REWARD_TANZANITE_IN_UCFX

RECYCLE_TRANSACTION_DELAY

Here is the delay for us to recycle those orphaned blocks in the boundary of eras and large epochs.

REWARD_EPOCH_COUNT

REWARD_EPOCH_COUNT needs to be larger than ANTICONE_PENALTY_UPPER_EPOCH_COUNT. If we cannot cache receipts of recent REWARD_EPOCH_COUNT epochs, the receipts will be loaded from db, which may lead to performance downgrade

TWO_YEAR_UNLOCK_TOKEN_COUNT_IN_CFX