cfxcore::db

Trait Cache

pub trait Cache<K, V> {
    // Required methods
    fn insert(&mut self, k: K, v: V) -> Option<V>;
    fn invalidate(&mut self, k: &K) -> Option<V>;
    fn get(&self, k: &K) -> Option<&V>;
}

A cache for arbitrary key-value pairs.

Required Methods

fn insert(&mut self, k: K, v: V) -> Option<V>

Insert an entry into the cache and get the old value.

fn invalidate(&mut self, k: &K) -> Option<V>

Invalidate an entry in the cache, getting the old value if it existed.

fn get(&self, k: &K) -> Option<&V>

Query the cache for a key’s associated value.