cfxcore

Module genesis_block

Structs

GenesisPosNodeInfo

GenesisPosState

Functions

default

genesis_block

test_net_version is used to update the genesis author so that after resetting, the chain of the older version will be discarded

load_file

load_secrets_file

register_transaction