cfxcore::genesis_block

Function genesis_block

pub fn genesis_block(
    storage_manager: &Arc<StorageManager>,
    genesis_accounts: HashMap<AddressWithSpace, U256>,
    test_net_version: Address,
    initial_difficulty: U256,
    machine: Arc<Machine>,
    need_to_execute: bool,
    genesis_chain_id: Option<u32>,
    initial_nodes: &Option<GenesisPosState>,
) -> Block

test_net_version is used to update the genesis author so that after resetting, the chain of the older version will be discarded