Module cfxcore::light_protocol

Re-exports

pub use query_service::QueryService;

Modules

query_service

Structs

Handler

Handler is responsible for maintaining peer meta-information and dispatching messages to the query and sync sub-handlers.

LightNodeConfiguration

Provider

Enums

Error

Constants

LIGHT_PROTOCOL_VERSION

LIGHT_PROTO_V1

The version to pass to Message for their lifetime declaration.

LIGHT_PROTO_V2