cfxcore::sync

Module request_manager

Modules

tx_handler

Structs

RequestHandler

RequestManager

When a header or block is requested by the RequestManager, it is ensured that if it’s not fully received, its hash exists in in_flight after every function call.

RequestMessage

SynchronizationPeerRequest

Traits

AsAny

Support to downcast trait to concrete request type.

Request

Trait of request message

Functions

try_get_block_hashes

Return block hashes in request if it’s requesting blocks. Return None otherwise.