cfxcore::sync::request_manager

Module tx_handler

Structs

InflightPendingTransactionContainer

InflightPendingTransactionItem

ReceivedTransactionContainer

SentTransactionContainer

This struct is not implemented as thread-safe since currently it is only used under protection of lock on SynchronizationState. Later we may refine the locking design to make it thread-safe.

TransactionCacheContainer