Module cfxcore::sync_parameters

Structs

FAILED_REQUEST_RESEND_WAIT

REQUEST_START_WAITING_TIME

Constants

BLOCK_PROPAGATION_DELAY

Measured block propagation delay in seconds. This will determine the conservative window when we measure confirmation risk internally in the consensus layer.

CATCH_UP_EPOCH_LAG_THRESHOLD

The threshold controlling whether a node is in catch-up mode. A node is in catch-up mode if its local best epoch number is CATCH_UP_EPOCH_LAG_THRESHOLD behind the median of the epoch numbers of peers.

DEFAULT_CHUNK_SIZE

LOCAL_BLOCK_INFO_QUERY_THRESHOLD

The threshold controlling whether we should query local_block_info in disk when requesting block header or block. If the difference between height of the block and current best height is less than LOCAL_BLOCK_INFO_QUERY_THRESHOLD, we can request block directly through network, otherwise we should check disk first.

MAX_BLOCKS_TO_SEND

The max number of blocks that are to be sent for compact block request.

MAX_EPOCHS_TO_SEND

The max number of epochs whose hashes are to be responded for request GetBlockHashesByEpoch

MAX_HEADERS_TO_SEND

The max number of headers that are to be sent for header block request.

MAX_PACKET_SIZE

REQUEST_TERMINAL_EPOCH_LAG_THRESHOLD

This threshold controlling whether a node should request missing terminals from peers when the node is in catch-up mode.