cfxcore::consensus

Module consensus_inner

Modules

confirmation_meter

consensus_executor

consensus_new_block_handler

Structs

ConsensusGraphInner

Implementation details of Eras, Timer chain and Checkpoints

ConsensusGraphNode

ConsensusGraphNodeData

ConsensusGraphNodeData contains all extra information of a block that will change as the consensus graph state evolves (e.g., pivot chain changes). Unlike the ConsensusGraphNode fields, fields in ConsensusGraphNodeData will only be available after the block is preactivated (after calling preactivate_block().

ConsensusInnerConfig

StateBlameInfo