cfxcore

Module db

Database utilities and definitions.

Enums

CacheUpdatePolicy

Modes for updating caches.

Constants

COL_BLAMED_HEADER_VERIFIED_ROOTS

Column for verified roots of blamed headers on light nodes

COL_BLOCKS

Column for Blocks.

COL_BLOCK_TRACES

Column for block traces

COL_EPOCH_NUMBER

Column for Epoch Sets

COL_HASH_BY_BLOCK_NUMBER

Column for block number index

COL_MISC

Column for miscellaneous items

COL_REWARD_BY_POS_EPOCH

Column for PoS interest reward info.

COL_TX_INDEX

Column for Transaction Index

NUM_COLUMNS

Number of columns in DB

Traits

Cache

A cache for arbitrary key-value pairs.

Key

Should be used to get database key associated with given value.

Readable

Should be used to read values from database.

Writable

Should be used to write value into database.