cfxcore::message

Trait MessageProtocolVersionBound

pub trait MessageProtocolVersionBound {
    // Required methods
    fn version_introduced(&self) -> ProtocolVersion;
    fn version_valid_till(&self) -> ProtocolVersion;
}

Required Methods

fn version_introduced(&self) -> ProtocolVersion

This message is introduced since this version.

fn version_valid_till(&self) -> ProtocolVersion

This message is valid until the specified version.

The return type is NOT defined as Option intentionally, because I’d like to make it impossible to keep a Message forever by default.

Whenever we bump a protocol version, always update the version_valid_till for each message.