Module cfxcore::sync

Modules

message

msg_sender

random

request_manager

utils

Structs

CatchUpCheckpointPhase

CatchUpFillBlockBodyPhase

CatchUpRecoverBlockHeaderFromDbPhase

CatchUpSyncBlockHeaderPhase

CatchUpSyncBlockPhase

LocalMessageTask

NormalSyncPhase

ProtocolConfiguration

StateSyncConfiguration

SyncGraphConfig

SyncGraphStatistics

SynchronizationGraph

SynchronizationGraphInner

SynchronizationGraphNode

SynchronizationPeerState

SynchronizationPhaseManager

SynchronizationProtocolHandler

SynchronizationService

SynchronizationState

Enums

Error

SyncHandlerWorkType

SyncPhaseType

Both Archive and Full node go through the following phases: CatchUpRecoverBlockHeaderFromDB –> CatchUpSyncBlockHeader –> CatchUpCheckpoint –> CatchUpFillBlockBody –> CatchUpSyncBlock –> Normal

Constants

CHECK_RPC_REQUEST_TIMER

SYNC_PROTO_V1

The version to pass to Message for their lifetime declaration.

SYNC_PROTO_V2

SYNC_PROTO_V3

Traits

SynchronizationPhaseTrait

Type Aliases

SharedSynchronizationGraph

SharedSynchronizationService